+++ unoptimized/include/constraints.hpp
@@ -10,37 +10,37 @@
 extern const Eigen::Matrix<double, 7, 1> ground;
 
 // Funkcje pomocnicze do mapowania stanu układu
-Eigen::Map<const Eigen::VectorXd> get_body_position(const Eigen::VectorXd& q, int id, const std::vector<int>& body_ids);
-Eigen::Map<const Eigen::VectorXd> get_body_rotation(const Eigen::VectorXd& q, int id, const std::vector<int>& body_ids);
+Eigen::Map<const Eigen::VectorXd> get_body_position(const Eigen::VectorXd& q, long int id, const std::vector<long int>& body_ids);
+Eigen::Map<const Eigen::VectorXd> get_body_rotation(const Eigen::VectorXd& q, long int id, const std::vector<long int>& body_ids);
 
 // Klasa bazowa dla ograniczeń
 class Constraint
 {
 public:
-    Constraint(int id, int body1_id, int body2_id);
+    Constraint(long int id, long int body1_id, long int body2_id);
 
     virtual std::shared_ptr<Constraint> clone() const = 0;
 
-    virtual Eigen::VectorXd ConstrainingFunctions(const Eigen::VectorXd& q, double t, const std::vector<int>& body_ids);
+    virtual Eigen::VectorXd ConstrainingFunctions(const Eigen::VectorXd& q, double t, const std::vector<long int>& body_ids);
     virtual double equations_number();
 
 protected:
-    int id;
-    int body1_id;
-    int body2_id;
+    long int id;
+    long int body1_id;
+    long int body2_id;
 };
 
 // Ograniczenie odległości między punktami na ciałach
 class DistanceConstraint : public Constraint
 {
 public:
-    DistanceConstraint(int id, int body1_id, int body2_id, const Eigen::Vector3d& body1_point,
+    DistanceConstraint(long int id, long int body1_id, long int body2_id, const Eigen::Vector3d& body1_point,
                        const Eigen::Vector3d& body2_point,
                        const Eigen::Vector3d& (*distance)(double t));
 
     std::shared_ptr<Constraint> clone() const override;
 
-    Eigen::VectorXd ConstrainingFunctions(const Eigen::VectorXd& q, double t, const std::vector<int>& body_ids) override;
+    Eigen::VectorXd ConstrainingFunctions(const Eigen::VectorXd& q, double t, const std::vector<long int>& body_ids) override;
     double equations_number() override;
 
 private:
@@ -53,11 +53,11 @@ private:
 class FixedParameterConstraint : public Constraint
 {
 public:
-    FixedParameterConstraint(int id, int body_id, int parameter_index);
+    FixedParameterConstraint(long int id, long int body_id, int parameter_index);
 
     std::shared_ptr<Constraint> clone() const override;
 
-    Eigen::VectorXd ConstrainingFunctions(const Eigen::VectorXd& q, double t, const std::vector<int>& body_ids) override;
+    Eigen::VectorXd ConstrainingFunctions(const Eigen::VectorXd& q, double t, const std::vector<long int>& body_ids) override;
     double equations_number() override;
 
 private:
@@ -68,11 +68,11 @@ private:
 class FixedOrientationConstraint : public Constraint
 {
 public:
-    FixedOrientationConstraint(int id, int body_id, const Eigen::Vector3d& orientation);
+    FixedOrientationConstraint(long int id, long int body_id, const Eigen::Vector3d& orientation);
 
     std::shared_ptr<Constraint> clone() const override;
 
-    Eigen::VectorXd ConstrainingFunctions(const Eigen::VectorXd& q, double t, const std::vector<int>& body_ids) override;
+    Eigen::VectorXd ConstrainingFunctions(const Eigen::VectorXd& q, double t, const std::vector<long int>& body_ids) override;
     double equations_number() override;
 
 private:
@@ -83,11 +83,11 @@ private:
 class FixedPositionConstraint : public Constraint
 {
 public:
-    FixedPositionConstraint(int id, int body_id, const Eigen::Vector3d& position);
+    FixedPositionConstraint(long int id, long int body_id, const Eigen::Vector3d& position);
 
     std::shared_ptr<Constraint> clone() const override;
 
-    Eigen::VectorXd ConstrainingFunctions(const Eigen::VectorXd& q, double t, const std::vector<int>& body_ids) override;
+    Eigen::VectorXd ConstrainingFunctions(const Eigen::VectorXd& q, double t, const std::vector<long int>& body_ids) override;
     double equations_number() override;
 
 private:
@@ -98,12 +98,12 @@ private:
 class BallJointConstraint : public Constraint
 {
 public:
-    BallJointConstraint(int id, int body1_id, int body2_id, const Eigen::Vector3d& body1_point,
+    BallJointConstraint(long int id, long int body1_id, long int body2_id, const Eigen::Vector3d& body1_point,
                         const Eigen::Vector3d& body2_point);
 
     std::shared_ptr<Constraint> clone() const override;
 
-    Eigen::VectorXd ConstrainingFunctions(const Eigen::VectorXd& q, double t, const std::vector<int>& body_ids) override;
+    Eigen::VectorXd ConstrainingFunctions(const Eigen::VectorXd& q, double t, const std::vector<long int>& body_ids) override;
     double equations_number() override;
 
 private:
