+++ unoptimized/src/bodies.cpp
@@ -2,7 +2,7 @@
 #include <cmath>
 #include <Eigen/Dense>
 
-Body::Body(int id, double x, double y, double z, double e0, double e1, double e2, double e3)
+Body::Body(long int id, double x, double y, double z, double e0, double e1, double e2, double e3)
     : id(id), q(7)
 {
     double norm = std::sqrt(e0*e0 + e1*e1 + e2*e2 + e3*e3);
@@ -23,7 +23,7 @@ const Eigen::VectorXd& Body::getPosition() const
     return q;
 }
 
-int Body::getId() const
+long int Body::getId() const
 {
     return id;
 }
